Flatten directory structure- rename tests to relate better to functionality.
