+++ Makefile
@@ -6,6 +6,6 @@ node_modules:
 .PHONY: test
 test: node_modules
 	./node_modules/.bin/mocha \
-		test/mocha/child.js \
-		test/mocha/integration.js \
-		test/mocha/parent.js
+		test/mocha/parentrunner.js \
+		test/childrunner.js \
+		test/bin/marionette-mocha.js
