+++ lib/childrunner.js
@@ -18,7 +18,7 @@ var MOCHA_BINARY = '/bin/_mocha';
  *
  * @param {Array} argv for process.
  */
-function Child(argv) {
+function ChildRunner(argv) {
   this.argv = argv;
 
   // path to actual mocha binary
@@ -26,7 +26,7 @@ function Child(argv) {
     fsPath.join(fsPath.dirname(require.resolve('mocha')), MOCHA_BINARY);
 }
 
-Child.prototype = {
+ChildRunner.prototype = {
   /**
    * mocha-json-proxy consumer instance.
    *
@@ -56,4 +56,4 @@ Child.prototype = {
   }
 };
 
-module.exports.Child = Child;
+module.exports.ChildRunner = ChildRunner;
