+++ lib/parentrunner.js
@@ -1,10 +1,10 @@
-var Child = require('./child').Child;
+var Child = require('./childrunner').ChildRunner;
 
-function Parent(argv) {
+function ParentRunner(argv) {
   this.argv = argv;
 }
 
-Parent.prototype = {
+ParentRunner.prototype = {
   /**
    * Runs the mocha tests with a given reporter.
    *
@@ -27,4 +27,4 @@ Parent.prototype = {
   }
 };
 
-module.exports.Parent = Parent;
+module.exports.ParentRunner = ParentRunner;
