+++ test/fixtures/with-helper.js
